refactor: split order_manager.py into multi-file module
- Create order_manager/ directory with logical separation of concerns
- Core functionality in core.py with OrderManager class
- Order placement methods in order_types.py
- Bracket order strategies in bracket_orders.py
- Position-related orders in position_orders.py
- Real-time tracking in tracking.py
- Type definitions in types.py
- Utility functions in utils.py
- Add protocols.py for proper mixin typing
- Convert relative imports to absolute imports
- Maintain backward compatibility - no API changes
- Fix all linting errors and type annotations

This refactoring improves code organization, maintainability, and
makes the codebase easier to navigate while preserving all existing
functionality.
